// playwright-e2e/models/ccd-events/cui-ccd-events/cui-ccd-events.ts
import CCDEvent from "../ccdEvent.ts";

type CuiCCDEvents = {
  CREATE_LIP_CLAIM: CCDEvent;
  CREATE_CLAIM_SPEC_AFTER_PAYMENT: CCDEvent;
  DEFENDANT_RESPONSE_CUI: CCDEvent;
  CLAIMANT_RESPONSE_CUI: CCDEvent;
  EVIDENCE_UPLOAD_APPLICANT: CCDEvent;
  EVIDENCE_UPLOAD_RESPONDENT: CCDEvent;
  TRIAL_READINESS: CCDEvent;
  REQUEST_FOR_RECONSIDERATION: CCDEvent;
  QUERY_MANAGEMENT_RAISE: CCDEvent;
  QUERY_MANAGEMENT_RESPOND: CCDEvent;
};

export default CuiCCDEvents;